requirements.txt
flet>=0.80.1
sqlcipher3-wheels>=0.5.0
cryptography>=42.0.0
fpdf2>=2.7.8
PyPDFForm>=1.4.21
Pillow>=10.0.0
reportlab>=4.0.0
pyzipper>=0.3.6

# Path resolution (Phase 4+)
platformdirs>=4.0.0

# AI / document ingestion (Phase 5.0)
llama-cpp-python>=0.2.85
llama-index>=0.10.0
llama-index-llms-llama-cpp>=0.1.0
sqlite-vec>=0.1.0
pypdf>=4.0.0
pdf2image>=1.17.0
# rapidocr-onnxruntime>=1.3.0  # No Python 3.13 wheel yet; skip until available.
#                               # App works without it; scanned PDF OCR silently skips.
requests>=2.31.0